requirements.txt
argparse>=1.4.0
getmac>=0.8.2
time>=1.0.0
GitPython>=3.0.5
